[FIX] move classes to their canonical files
