+++ account_operating_unit/models/__init__.py
@@ -1,8 +1,9 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
 from . import res_company
-from . import account_bank_statement
+from . import account_bank_statement_line
 from . import account_journal
 from . import account_move
+from . import account_move_line
 from . import account_partial_reconcile
 from . import account_payment
